2026/ai internship/cognitive routing and rag/requirements.txt
langchain
langgraph
faiss-cpu
sentence-transformers
numpy